Add support for two new BBCode tags - url_map and heading
"heading" is to emulate h1-h6 tags. It's not very interesting. I had
trouble getting the template function to work with the level, so I had
to do that manually.

The more interesting tag is "url_map". You pass it an reference to an
identifier in a map of IDs to URLs, which is stored in the new config
folder system. The idea is to ensure that URLs to certain things are
kept up to date more easily.