Replace mustache template engine with handlebars template engine
I'm not really sold on the engine I've selected, but it's the only one
with something of a reasonable API. In order to shield the rest of the
codebase from further changes to the template engine, I've implemented
an interface and adapter for the engine.

The reason for moving to handlebars is to get support for helpers.
Initially, I need support for formatting dates, as all dates are stored
as Unix timestamps. I couldn't achieve that natively with mustache
templates.

In theory, all existing templates should continue to function without
any changes. The only necessary change at this stage is to update the
file extensions for templates from "mustache" to "hbs".